requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0